+++ scilpy/denoise/asym_averaging.py
