+++ scilpy/denoise/bilateral_filtering.py
